--- a/src/SFML/Audio/AudioDevice.cpp
+++ b/src/SFML/Audio/AudioDevice.cpp
@@ -29,31 +29,57 @@
 
 #include <SFML/System/Err.hpp>
 
+#include <miniaudio.h>
+
 #include <algorithm>
+#include <memory>
+#include <optional>
 #include <ostream>
 
 
 namespace sf::priv
 {
 ////////////////////////////////////////////////////////////
-AudioDevice::AudioDevice()
+struct AudioDevice::ListenerProperties
+{
+    float          volume{100.f};
+    sf::Vector3f   position{0, 0, 0};
+    sf::Vector3f   direction{0, 0, -1};
+    sf::Vector3f   velocity{0, 0, 0};
+    Listener::Cone cone{sf::degrees(360), sf::degrees(360), 1};
+    sf::Vector3f   upVector{0, 1, 0};
+};
+
+
+////////////////////////////////////////////////////////////
+struct AudioDevice::Impl
+{
+    std::optional<ma_log>     log;            //!< The miniaudio log
+    std::optional<ma_context> context;        //!< The miniaudio context
+    std::optional<ma_device>  playbackDevice; //!< The miniaudio playback device
+    std::optional<ma_engine>  engine;         //!< The miniaudio engine (used for effects and spatialisation)
+};
+
+
+////////////////////////////////////////////////////////////
+AudioDevice::AudioDevice() : m_impl(std::make_unique<Impl>())
 {
     // Ensure we only ever have a single AudioDevice instance
     assert(getInstance() == nullptr);
     getInstance() = this;
 
     // Create the log
-    m_log.emplace();
+    m_impl->log.emplace();
 
-    if (const auto result = ma_log_init(nullptr, &*m_log); result != MA_SUCCESS)
+    if (const auto result = ma_log_init(nullptr, &*m_impl->log); result != MA_SUCCESS)
     {
-        m_log.reset();
+        m_impl->log.reset();
         err() << "Failed to initialize the audio log: " << ma_result_description(result) << std::endl;
         return;
     }
 
     // Register our logging callback to output any warning/error messages
-    if (const auto result = ma_log_register_callback(&*m_log,
+    if (const auto result = ma_log_register_callback(&*m_impl->log,
                                                      ma_log_callback_init(
                                                          [](void*, ma_uint32 level, const char* message)
                                                          {
@@ -66,14 +92,14 @@ AudioDevice::AudioDevice()
         err() << "Failed to register audio log callback: " << ma_result_description(result) << std::endl;
 
     // Create the context
-    m_context.emplace();
+    m_impl->context.emplace();
 
     auto contextConfig = ma_context_config_init();
-    contextConfig.pLog = &*m_log;
+    contextConfig.pLog = &*m_impl->log;
 
-    if (const auto result = ma_context_init(nullptr, 0, &contextConfig, &*m_context); result != MA_SUCCESS)
+    if (const auto result = ma_context_init(nullptr, 0, &contextConfig, &*m_impl->context); result != MA_SUCCESS)
     {
-        m_context.reset();
+        m_impl->context.reset();
         err() << "Failed to initialize the audio context: " << ma_result_description(result) << std::endl;
         return;
     }
@@ -81,7 +107,7 @@ AudioDevice::AudioDevice()
     // Count the playback devices
     ma_uint32 deviceCount = 0;
 
-    if (const auto result = ma_context_get_devices(&*m_context, nullptr, &deviceCount, nullptr, nullptr);
+    if (const auto result = ma_context_get_devices(&*m_impl->context, nullptr, &deviceCount, nullptr, nullptr);
         result != MA_SUCCESS)
     {
         err() << "Failed to get audio playback devices: " << ma_result_description(result) << std::endl;
@@ -96,16 +122,16 @@ AudioDevice::AudioDevice()
     }
 
     // Create the playback device
-    m_playbackDevice.emplace();
+    m_impl->playbackDevice.emplace();
 
     auto playbackDeviceConfig         = ma_device_config_init(ma_device_type_playback);
     playbackDeviceConfig.dataCallback = [](ma_device* device, void* output, const void*, ma_uint32 frameCount)
     {
         auto& audioDevice = *static_cast<AudioDevice*>(device->pUserData);
 
-        if (audioDevice.m_engine)
+        if (audioDevice.m_impl->engine)
         {
-            if (const auto result = ma_engine_read_pcm_frames(&*audioDevice.m_engine, output, frameCount, nullptr);
+            if (const auto result = ma_engine_read_pcm_frames(&*audioDevice.m_impl->engine, output, frameCount, nullptr);
                 result != MA_SUCCESS)
                 err() << "Failed to read PCM frames from audio engine: " << ma_result_description(result) << std::endl;
         }
@@ -113,50 +139,54 @@ AudioDevice::AudioDevice()
     playbackDeviceConfig.pUserData       = this;
     playbackDeviceConfig.playback.format = ma_format_f32;
 
-    if (const auto result = ma_device_init(&*m_context, &playbackDeviceConfig, &*m_playbackDevice); result != MA_SUCCESS)
+    if (const auto result = ma_device_init(&*m_impl->context, &playbackDeviceConfig, &*m_impl->playbackDevice);
+        result != MA_SUCCESS)
     {
-        m_playbackDevice.reset();
+        m_impl->playbackDevice.reset();
         err() << "Failed to initialize the audio playback device: " << ma_result_description(result) << std::endl;
         return;
     }
 
     // Create the engine
     auto engineConfig          = ma_engine_config_init();
-    engineConfig.pContext      = &*m_context;
-    engineConfig.pDevice       = &*m_playbackDevice;
+    engineConfig.pContext      = &*m_impl->context;
+    engineConfig.pDevice       = &*m_impl->playbackDevice;
     engineConfig.listenerCount = 1;
 
-    m_engine.emplace();
+    m_impl->engine.emplace();
 
-    if (const auto result = ma_engine_init(&engineConfig, &*m_engine); result != MA_SUCCESS)
+    if (const auto result = ma_engine_init(&engineConfig, &*m_impl->engine); result != MA_SUCCESS)
     {
-        m_engine.reset();
+        m_impl->engine.reset();
         err() << "Failed to initialize the audio engine: " << ma_result_description(result) << std::endl;
         return;
     }
 
     // Set master volume, position, velocity, cone and world up vector
-    if (const auto result = ma_device_set_master_volume(ma_engine_get_device(&*m_engine),
+    if (const auto result = ma_device_set_master_volume(ma_engine_get_device(&*m_impl->engine),
                                                         getListenerProperties().volume * 0.01f);
         result != MA_SUCCESS)
         err() << "Failed to set audio device master volume: " << ma_result_description(result) << std::endl;
 
-    ma_engine_listener_set_position(&*m_engine,
+    ma_engine_listener_set_position(&*m_impl->engine,
                                     0,
                                     getListenerProperties().position.x,
                                     getListenerProperties().position.y,
                                     getListenerProperties().position.z);
-    ma_engine_listener_set_velocity(&*m_engine,
+
+    ma_engine_listener_set_velocity(&*m_impl->engine,
                                     0,
                                     getListenerProperties().velocity.x,
                                     getListenerProperties().velocity.y,
                                     getListenerProperties().velocity.z);
-    ma_engine_listener_set_cone(&*m_engine,
+
+    ma_engine_listener_set_cone(&*m_impl->engine,
                                 0,
                                 getListenerProperties().cone.innerAngle.asRadians(),
                                 getListenerProperties().cone.outerAngle.asRadians(),
                                 getListenerProperties().cone.outerGain);
-    ma_engine_listener_set_world_up(&*m_engine,
+
+    ma_engine_listener_set_world_up(&*m_impl->engine,
                                     0,
                                     getListenerProperties().upVector.x,
                                     getListenerProperties().upVector.y,
@@ -168,20 +198,20 @@ AudioDevice::AudioDevice()
 AudioDevice::~AudioDevice()
 {
     // Destroy the engine
-    if (m_engine)
-        ma_engine_uninit(&*m_engine);
+    if (m_impl->engine)
+        ma_engine_uninit(&*m_impl->engine);
 
     // Destroy the playback device
-    if (m_playbackDevice)
-        ma_device_uninit(&*m_playbackDevice);
+    if (m_impl->playbackDevice)
+        ma_device_uninit(&*m_impl->playbackDevice);
 
     // Destroy the context
-    if (m_context)
-        ma_context_uninit(&*m_context);
+    if (m_impl->context)
+        ma_context_uninit(&*m_impl->context);
 
     // Destroy the log
-    if (m_log)
-        ma_log_uninit(&*m_log);
+    if (m_impl->log)
+        ma_log_uninit(&*m_impl->log);
 
     // Ensure we only ever have a single AudioDevice instance
     assert(getInstance() != nullptr);
@@ -194,8 +224,8 @@ ma_engine* AudioDevice::getEngine()
 {
     auto* instance = getInstance();
 
-    if (instance && instance->m_engine)
-        return &*instance->m_engine;
+    if (instance && instance->m_impl->engine)
+        return &*instance->m_impl->engine;
 
     return nullptr;
 }
@@ -209,10 +239,10 @@ void AudioDevice::setGlobalVolume(float volume)
 
     auto* instance = getInstance();
 
-    if (!instance || !instance->m_engine)
+    if (!instance || !instance->m_impl->engine)
         return;
 
-    if (const auto result = ma_device_set_master_volume(ma_engine_get_device(&*instance->m_engine), volume * 0.01f);
+    if (const auto result = ma_device_set_master_volume(ma_engine_get_device(&*instance->m_impl->engine), volume * 0.01f);
         result != MA_SUCCESS)
         err() << "Failed to set audio device master volume: " << ma_result_description(result) << std::endl;
 }
@@ -233,10 +263,10 @@ void AudioDevice::setPosition(const Vector3f& position)
 
     auto* instance = getInstance();
 
-    if (!instance || !instance->m_engine)
+    if (!instance || !instance->m_impl->engine)
         return;
 
-    ma_engine_listener_set_position(&*instance->m_engine, 0, position.x, position.y, position.z);
+    ma_engine_listener_set_position(&*instance->m_impl->engine, 0, position.x, position.y, position.z);
 }
 
 
@@ -255,10 +285,10 @@ void AudioDevice::setDirection(const Vector3f& direction)
 
     auto* instance = getInstance();
 
-    if (!instance || !instance->m_engine)
+    if (!instance || !instance->m_impl->engine)
         return;
 
-    ma_engine_listener_set_direction(&*instance->m_engine, 0, direction.x, direction.y, direction.z);
+    ma_engine_listener_set_direction(&*instance->m_impl->engine, 0, direction.x, direction.y, direction.z);
 }
 
 
@@ -277,10 +307,10 @@ void AudioDevice::setVelocity(const Vector3f& velocity)
 
     auto* instance = getInstance();
 
-    if (!instance || !instance->m_engine)
+    if (!instance || !instance->m_impl->engine)
         return;
 
-    ma_engine_listener_set_velocity(&*instance->m_engine, 0, velocity.x, velocity.y, velocity.z);
+    ma_engine_listener_set_velocity(&*instance->m_impl->engine, 0, velocity.x, velocity.y, velocity.z);
 }
 
 
@@ -299,10 +329,10 @@ void AudioDevice::setCone(const Listener::Cone& cone)
 
     auto* instance = getInstance();
 
-    if (!instance || !instance->m_engine)
+    if (!instance || !instance->m_impl->engine)
         return;
 
-    ma_engine_listener_set_cone(&*instance->m_engine,
+    ma_engine_listener_set_cone(&*instance->m_impl->engine,
                                 0,
                                 std::clamp(cone.innerAngle.asRadians(), 0.f, sf::degrees(360).asRadians()),
                                 std::clamp(cone.outerAngle.asRadians(), 0.f, sf::degrees(360).asRadians()),
@@ -325,10 +355,10 @@ void AudioDevice::setUpVector(const Vector3f& upVector)
 
     auto* instance = getInstance();
 
-    if (!instance || !instance->m_engine)
+    if (!instance || !instance->m_impl->engine)
         return;
 
-    ma_engine_listener_set_world_up(&*instance->m_engine, 0, upVector.x, upVector.y, upVector.z);
+    ma_engine_listener_set_world_up(&*instance->m_impl->engine, 0, upVector.x, upVector.y, upVector.z);
 }