The header-only version of miniaudio is almost 4MB. This means that *every* translation unit that includes `miniaudio.h` basically copy-pastes 4MB inside itself, even though all translation units but one need the implementation.

Split version is available directly from the miniaudio repo:
https://github.com/mackron/miniaudio/tree/master/extras/miniaudio_split