@vittorioromeo 2024-05-07
@ChrisThrasher: what would be the proper way of compiling and linking against `miniaudio.c`?