@ChrisThrasher 2024-05-07
I measured compiling `sfml-audio` on `master` and this PR and build times were identical. It seems like modern preprocessors are really good at churning through all those megabytes of code.

`master`
```
$ hyperfine 'ninja clean && ninja sfml-audio'
Benchmark 1: ninja clean && ninja sfml-audio
  Time (mean ± σ):      1.981 s ±  0.017 s    [User: 11.487 s, System: 1.522 s]
  Range (min … max):    1.958 s …  2.009 s    10 runs
```

#2982
```
$ hyperfine 'ninja clean && ninja sfml-audio'
Benchmark 1: ninja clean && ninja sfml-audio
  Time (mean ± σ):      1.981 s ±  0.044 s    [User: 11.390 s, System: 1.500 s]
  Range (min … max):    1.901 s …  2.056 s    10 runs
```