@ChrisThrasher 2024-05-07
> @ChrisThrasher: what would be the proper way of compiling and linking against `miniaudio.c`?

What you did `#include`ing the source file seems fine. A bit unconventional but it looks like it will work.

Something to keep in mind is that the way we run clang-tidy relies on analyzing all files that appear in compile_commands.json. If we add a 3rd party source file to the build (as in we actually tell CMake about it) then it will appear in that file and thus get analyzed. No 3rd party source file is going to pass all of our clang-tidy checks so that will cause CI to break. What you proposed seems to avoid that issue since the build system has no idea we added a new source file.