@vittorioromeo 2024-05-07
> I measured compiling `sfml-audio` on `master` and this PR and build times were identical. It seems like modern preprocessors are really good at churning through all those megabytes of code.
> 
> `master`
> 
> ```
> $ hyperfine 'ninja clean && ninja sfml-audio'
> Benchmark 1: ninja clean && ninja sfml-audio
>   Time (mean ± σ):      1.981 s ±  0.017 s    [User: 11.487 s, System: 1.522 s]
>   Range (min … max):    1.958 s …  2.009 s    10 runs
> ```
> 
> #2982
> 
> ```
> $ hyperfine 'ninja clean && ninja sfml-audio'
> Benchmark 1: ninja clean && ninja sfml-audio
>   Time (mean ± σ):      1.981 s ±  0.044 s    [User: 11.390 s, System: 1.500 s]
>   Range (min … max):    1.901 s …  2.056 s    10 runs
> ```

I see a slight improvement with this PR, but YMMV. I also think you should benchmark tests and examples as well, to measure the impact on the TUs that might include headers that include `miniaudio.h`:

```
this PR 
Benchmark 1: ninja clean && ninja sfml-audio
  Time (mean ± σ):      1.809 s ±  0.020 s    [User: 5.190 s, System: 1.812 s]
  Range (min … max):    1.788 s …  1.831 s    4 runs

Benchmark 1: ninja clean && ninja
  Time (mean ± σ):      4.589 s ±  0.047 s    [User: 85.597 s, System: 21.479 s]
  Range (min … max):    4.533 s …  4.632 s    4 runs



master
Benchmark 1: ninja clean && ninja sfml-audio
  Time (mean ± σ):      1.889 s ±  0.040 s    [User: 5.218 s, System: 1.952 s]
  Range (min … max):    1.851 s …  1.944 s    4 runs

Benchmark 1: ninja clean && ninja
  Time (mean ± σ):      4.645 s ±  0.080 s    [User: 85.421 s, System: 21.772 s]
  Range (min … max):    4.535 s …  4.721 s    4 runs
```