@ChrisThrasher 2024-05-07
> to measure the impact on the TUs that might include headers that include miniaudio.h:

This header is not included in any public headers which is why I ignored tests and examples.