@vittorioromeo 2024-05-07
> > to measure the impact on the TUs that might include headers that include miniaudio.h:
> 
> miniaudio.h is not included in any public SFML headers which is why I ignored tests and examples.

I apologize, you are correct. I didn't realize `AudioDevice.hpp` was under `src/`.