@eXpl0it3r 2024-05-07
> The header-only version of miniaudio is almost 4MB. This means that _every_ translation unit that includes `miniaudio.h` basically copy-pastes 4MB inside itself, even though all translation units but one need the implementation.

How does this PR change anything about this situation? Since you include a C file, isn't the outcome exactly the same?

Also, I'm against doing weird things like including source files to shave off 0.1s build time on a specific system. The analyzers are also not happy for good reasons, and I wouldn't want to silence them just for this.