@binary1248 2024-05-07
Besides the points already mentioned above, bear in mind that in the case of miniaudio, the maintainer explicitly mentions in the split files readme that they are derived from the single header file and not the other way around. As such, the miniaudio single header is officially the preferred way of using miniaudio if a project has the choice of choosing between both methods.

The reason why this change basically doesn't change build times is because including the single header without defining `MINIAUDIO_IMPLEMENTATION` is equivalent to including the split header. After preprocessing there is basically no difference for the compiler. Sure, the preprocessor has to skip most of the file, but modern operating systems would have already pulled the entire contents into RAM after the first time it was read so subsequent reads will run as fast as a `memcpy`. Because it is a common usage pattern, I'm sure the preprocessor also has a bunch of optimizations to skip over huge blocks of code that aren't active because of a missing `#define` looking for the next `#else` or `#endif`.

Whoever looks for a performance benefit to this change will not find one, because it practically doesn't exist unless you are using an operating system and compiler from the 90s.