@vittorioromeo 2024-05-07
New results measured with ClangBuildAnalyzer:

```
#
#
# WITH PR

623 ms: C:/OHWorkspace/SFML/extlibs/headers/miniaudio/miniaudio.h (included 7 times, avg 89 ms), included via:
  4x: <direct include>
  3x: MiniaudioUtils.hpp 

#
#
# WITHOUT PR

674 ms: C:/OHWorkspace/SFML/extlibs/headers/miniaudio/miniaudio.h (included 12 times, avg 56 ms), included via:
  5x: AudioDevice.hpp 
  4x: <direct include>
  3x: MiniaudioUtils.hpp 
```

---

> How does this PR change anything about this situation? Since you include a C file, isn't the outcome exactly the same?

@eXpl0it3r: the `.c` file is included only in `Miniaudio.cpp`, while before every inclusion of `miniaudio.h` would copy-paste the implementation code as well.

---

> Also, I'm against doing weird things like including source files to shave off 0.1s build time on a specific system. 

@eXpl0it3r: I am happy to avoid the `#include <miniaudio.c>` and add the file to `CMakeLists` as part of `sfml-audio`'s sources.

---

> the maintainer explicitly mentions in the split files readme that they are derived from the single header file and not the other way around. As such, the miniaudio single header is officially the preferred way of using miniaudio if a project has the choice of choosing between both methods.

@binary1248: I had seen that, and verified that the split version was up-to-date. Considering our `extlibs` update workflow is manual, I don't see it as a big deal. E.g. we were behind on stb_image quite a bit before #2981.

---

> The reason why this change basically doesn't change build times [...]

@binary1248: I did see a difference in my benchmarks using `clang++`. 
https://github.com/SFML/SFML/pull/2982#issuecomment-2097166088

--- 

I've also added extra improvements to make this PR more useful.